+++ R/example_phenos.R
@@ -8,6 +8,6 @@
 #' @importFrom data.table data.table
 #' @examples
 #' phenos <- example_phenos()
-example_phenos <- function(i=seq_len(10)){
-  data.table::data.table(hpo_id=get_hpo()$id[i])
+example_phenos <- function(i=seq(10)){
+  data.table::data.table(hpo_id=grep("^HP:",get_hpo()$id, value = TRUE)[i])
 }