+++ R/get_hpo.R
@@ -9,21 +9,18 @@
 #'  (due to an expansion of the HPO).
 #' @source \href{https://bioportal.bioontology.org/ontologies/HP}{BioPortal}
 #' \code{
-#' save_path <- file.path(tempdir(),"hpo.rds")
-#' URL <- paste0(
-#'   "https://data.bioontology.org/ontologies/HP/submissions/602/",
-#'   "download?apikey=8b5b7825-538d-40e0-9e9e-5ab9274a9aeb"
-#' )
-#' tmp <- tempfile(fileext = "hpo.obo")
-#' utils::download.file(URL, tmp)
-#' hpo <- ontologyIndex::get_OBO(tmp, extract_tags = "everything")
-#' saveRDS(hpo,save_path)
+#' f <- HPOExplorer:::get_data(file = "hp-base.obo",
+#'                             repo = "obophenotype/human-phenotype-ontology",
+#'                             tag="latest")
+#' hpo <- ontologyIndex::get_OBO(f, extract_tags = "everything")
 #' ### Fix non-ASCII characters in metadata ####
 #' func <- function(v){
 #'   Encoding(v) <- "latin1"
 #'   iconv(v, "latin1", "UTF-8")
 #' }
 #' attributes(hpo)$version <- func(attributes(hpo)$version)
+#' save_path <- file.path(tempdir(),"hpo.rds")
+#' saveRDS(hpo,save_path)
 #' piggyback::pb_upload(file = save_path,
 #'                      repo = "neurogenomics/HPOExplorer",
 #'                      overwrite = TRUE,